pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "flask>=3.1.2",
    "moviepy>=2.2.1",
    "pillow>=11.3.0",
]
